+++ clases/Reportes.php
@@ -147,6 +147,7 @@ class Reportes {
   }
 
   public function eliminarImagen() {
+    if(!$this->imagen) return;
     $existeArchivo = file_exists(CARPETA_IMAGEN . $this->imagen);
     if($existeArchivo) {
       unlink(CARPETA_IMAGEN . $this->imagen);
